This runs for multibit input bus
Notes:
Ngspice does not support buses. At least characeters '<>'
Can not be used in node names. As a quick fix, Buses in Python
are still '<>', but they are converted to underscores in testbench.

[ToDo]
Decide how the output 'ionemes' of buses should be handled.
Currently I explicitly set them to ['Z_1_', 'Z_0_']